avanze 1
04_11_2022
Que aprendi?
No se XD
metodos,variables,bucles
