+++ nbproject/private/private.xml
@@ -3,11 +3,13 @@
     <editor-bookmarks xmlns="http://www.netbeans.org/ns/editor-bookmarks/2" lastBookmarkId="0"/>
     <open-files xmlns="http://www.netbeans.org/ns/projectui-open-files/2">
         <group>
-            <file>file:/C:/Users/NOT%20FOUND%20404/Documents/NetBeansProjects/JavaApplication1/src/Modulo3/estruturasdeControl.java</file>
-            <file>file:/C:/Users/NOT%20FOUND%20404/Documents/NetBeansProjects/JavaApplication1/src/Practicas/preciodeVenta.java</file>
-            <file>file:/C:/Users/NOT%20FOUND%20404/Documents/NetBeansProjects/JavaApplication1/src/Modulo3/esPar.java</file>
-            <file>file:/C:/Users/NOT%20FOUND%20404/Documents/NetBeansProjects/JavaApplication1/src/Modulo3/esVocal.java</file>
-            <file>file:/C:/Users/NOT%20FOUND%20404/Documents/NetBeansProjects/JavaApplication1/src/Proyectos/numero1.java</file>
+            <file>file:/C:/Users/NOT%20FOUND%20404/Documents/NetBeansProjects/MrJava/mrJava/src/Modulo3/bucleWhile.java</file>
+            <file>file:/C:/Users/NOT%20FOUND%20404/Documents/NetBeansProjects/MrJava/mrJava/src/Modulo3/esPar.java</file>
+            <file>file:/C:/Users/NOT%20FOUND%20404/Documents/NetBeansProjects/MrJava/mrJava/src/Modulo3/doWhileMultiplicacion.java</file>
+            <file>file:/C:/Users/NOT%20FOUND%20404/Documents/NetBeansProjects/MrJava/mrJava/src/Modulo3/swicthDias.java</file>
+            <file>file:/C:/Users/NOT%20FOUND%20404/Documents/NetBeansProjects/MrJava/mrJava/src/Modulo3/esVocal.java</file>
+            <file>file:/C:/Users/NOT%20FOUND%20404/Documents/NetBeansProjects/MrJava/mrJava/src/Modulo3/whileTablaMultiplicar.java</file>
+            <file>file:/C:/Users/NOT%20FOUND%20404/Documents/NetBeansProjects/MrJava/mrJava/src/Modulo3/doWhile.java</file>
         </group>
     </open-files>
 </project-private>
