+++ nbproject/project.properties
@@ -71,7 +71,7 @@ jlink.additionalmodules=
 jlink.additionalparam=
 jlink.launcher=true
 jlink.launcher.name=JavaApplication1
-main.class=Proyectos.numero1
+main.class=Variables.numero1
 manifest.file=manifest.mf
 meta.inf.dir=${src.dir}/META-INF
 mkdist.disabled=false
