--- src/Modulo3/estruturasdeControl.java
@@ -1,19 +1,0 @@
-package Modulo3;
-
-import java.util.Scanner;
-
-public class estruturasdeControl {
-
-    public static void main(String[] args) {
-        //If else
-        Scanner in = new Scanner(System.in);
-        if (true) {
-            System.out.println("Se cumplio la condición");
-        } else {
-
-            System.out.println("No se cumplio la condición");
-        }
-
-    }
-
-}
