+++ src/Modulo3estructurasdeControl/bucleWhile.java
@@ -1,0 +1,15 @@
+package Modulo3estructurasdeControl;
+
+public class bucleWhile {
+
+    public static void main(String[] args) {
+
+        //Estructura Repetitiva Mientras
+        System.out.print("Numero: ");
+        int num = 0;
+        while (num <= 10) {
+            System.out.println("El numero es : " + num);
+            num++;
+        }
+    }
+}
