+++ src/Modulo3estructurasdeControl/estruturasdeControl.java
@@ -1,0 +1,19 @@
+package Modulo3estructurasdeControl;
+
+import java.util.Scanner;
+
+public class estruturasdeControl {
+
+    public static void main(String[] args) {
+        //If else
+        Scanner in = new Scanner(System.in);
+        if (true) {
+            System.out.println("Se cumplio la condición");
+        } else {
+
+            System.out.println("No se cumplio la condición");
+        }
+        System.out.println("Adios");
+    }
+
+}
