+++ src/Modulo3estructurasdeControl/in.java
@@ -1,0 +1,18 @@
+
+package Modulo3estructurasdeControl;
+
+import java.util.Scanner;
+
+/**
+ *
+ * @author NOT FOUND 404
+ */
+class in {
+
+    static void nextInt() {
+        throw new UnsupportedOperationException("Not supported yet."); // Generated from nbfs://nbhost/SystemFileSystem/Templates/Classes/Code/GeneratedMethodBody
+    }
+
+    Scanner in = new Scanner(System.in);
+
+}
