--- src/Proyectos/posIncremento_posDecremento.java
@@ -1,35 +1,0 @@
-package Proyectos;
-
-/**
- *
- * @author NOT FOUND 404
- */
-public class posIncremento_posDecremento {
-
-    public static void main(String[] args) {
-
-        //Incremento c++
-        int c = 0;
-        c++;
-        c++;
-        System.out.println(c);
-        //Decremento
-        c--;
-        c--;
-        System.out.println(c);
-
-        //PostIncremento
-        //0 y c++ 1
-        System.out.println(c++);
-        System.out.println(c++);
-        System.out.println(c);
-
-        //PreIncremento
-        //0 y c++ 1
-        System.out.println(++c);
-        System.out.println(++c);
-        System.out.println(c);
-
-    }
-
-}
